requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.129.0
fonttools==4.61.1
h11==0.16.0
idna==3.11
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pandas==3.0.0
patsy==1.0.2
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
starlette==0.52.1
statsmodels==0.14.6
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
xgboost==3.2.0
